feat: 🎸 config typescript and ts-node-dev in project
I add typescript configuration to facilitate the developement proccess
and add ts-node-dev to develop in real-time
