+++ .gitignore
@@ -1,1 +1,8 @@
 node_modules/
+
+package-*.json
+package.json.*
+yarn.lock
+yarn-*.log
+
+dist
