+++ Procfile
@@ -1,1 +1,1 @@
-web: node src/server.js
+web: npm run build && node dist/server.js
