+++ package.json
@@ -5,8 +5,9 @@
   "main": "src/server.js",
   "scripts": {
     "commit": "npx git-cz",
-    "dev": "nodemon --ignore 'node_modules/*' src/server.js",
-    "format": "npx standard --fix"
+    "dev": "ts-node-dev --respawn --transpileOnly --ignore-watch node_modules src/server.ts",
+    "format": "npx standard --fix",
+    "build": "tsc"
   },
   "keywords": [],
   "author": "Mikael Rolim de Aquino <mikarg9@gmail.com>",
@@ -19,6 +20,13 @@
     "request": "^2.88.2"
   },
   "devDependencies": {
-    "nodemon": "^2.0.4"
+    "@types/cheerio": "^0.22.18",
+    "@types/cors": "^2.8.6",
+    "@types/express": "^4.17.6",
+    "@types/jsdom": "^16.2.3",
+    "@types/node": "^14.0.14",
+    "@types/request": "^2.48.5",
+    "ts-node-dev": "^1.0.0-pre.50",
+    "typescript": "^3.9.6"
   }
 }
