+++ src/server.ts
@@ -1,0 +1,15 @@
+import express  from 'express'
+import cors  from 'cors'
+
+import routes  from './routes'
+
+const app = express()
+
+const port = process.env.PORT || 3333
+
+app.use(cors())
+app.use(express.json())
+
+app.use(routes)
+
+app.listen(port, () => console.log('Server working on localhost:' + port))
